Make gem ready
In this commit, we've moved files that were previously in the main
causes repository and made them free from any causes-specific
dependencies. That meant...

- Introduce "Cat" model for testing, instead of using the "User" model
  as we did in the causes repo.
- Add sqlite schema for new "Cat" model.
- Roll our own version of `#pluralize` (to avoid adding a dependency on
  ActionView)

We are unsure if it would be best to roll our own `#pluralize` or to
bring in ActionView, but for the time being, we decided to roll our own.

This is likely compatible with older versions of ActiveRecord and
ActiveSupport, but for the time being we decided to start where we are
at in the causes project, which is 4.0+. A future commit might see the
addition of more backwards compatibility.

Change-Id: Iab4d8baca386b4d85b7b89ccc214699d0b2774c8
Reviewed-on: http://gerrit.causes.com/36988
Reviewed-by: Henric Trotzig <henric.trotzig@causes.com>
Tested-by: Henric Trotzig <henric.trotzig@causes.com>