Require 'rspec' inside make_database_queries.rb
We ran into issues using the gem in the Causes project where we got an
error on this line:

    require 'db_query_matchers'

We believe that requiring 'rspec' in the matcher file will fix that
issue. At least that made it work when referencing the db_query_matchers
gem locally on the file system.

Change-Id: Ib0fe839486ef8f3fdb7cdb779bc1bbd2a141b8ff