Fix Bundler auto-requiring
When using this gem with an app that uses Bundler, there will likely be
a line in the application config like (Rails example):

  Bundler.require(*Rails.groups)

This will auto-require all of the gems in the Gemfile based on the name
of the gem itself. However, since this gem uses dashes for its name, but
the file in lib uses underscores, this auto-requiring didn't work which
means that you would likely need to do something like:

  gem 'db-query-matchers', require: 'db_query_matchers'

To improve this, I am adding a file that matches the name of the gem
that requires the other file. This will allow the Bundler.require call
to do its thing without adding the explicit require option to the
Gemfile.

Change-Id: I7f1e34564dddd7662bc159275333395f8f1d4908
Reviewed-on: http://gerrit.causes.com/39247
Tested-by: Joe Lencioni <joe.lencioni@brigade.com>
Reviewed-by: Ryan Fitzgerald <ryan.fitzgerald@brigade.com>