Add configuration to ignore certain queries
In some cases, queries may happen that we don't want to be counted by
the `make_database_queries` matcher. For example, in Rails you might get
`SHOW TABLE LIKE 'foo'` queries. These meta queries might not always be
relevant. To ignore these types of queries, you can now specify an
`ignores` configuration to DBQueryMatchers. This configuration should
consist of an array of regular expressions that when matched for a
query, will cause the query not to be counted.

Change-Id: I8f60bf2b0273be90aede52c65481a9bb75b353b7
Reviewed-on: http://gerrit.causes.com/40189
Reviewed-by: Joe Lencioni <joe.lencioni@brigade.com>
Tested-by: Henric Trotzig <henric.trotzig@brigade.com>