Add .projections.json configuration file
Editor plugins, like @tpope's vim-projectionist, can make use of this
configuration file to enable easy jumping between "alternate" files,
among other features. For example, with this configuration and
vim-projectionist installed, you can now be editing a matcher and
execute `:A` to jump to that matcher's spec file. This will save some
time when working with this project.

https://github.com/tpope/vim-projectionist

Change-Id: I4cc061d08fe7fdeeb676bf8e5611135cb37e7bf1
Reviewed-on: http://gerrit.causes.com/45431
Reviewed-by: Joe Lencioni <joe.lencioni@brigade.com>
Tested-by: Joe Lencioni <joe.lencioni@brigade.com>