Add support for a generic matching clause
The reason for supporting both regular expressions and strings is that
we want to be able to be both flexible in my matching (hence the
regexp), and to quickly and easily match particular exact queries
generated by for example an ActiveRecord scope.

This allows you to do: expect { request }.to_not
make_database_queries(matching: User.deleted.to_sql).

Change-Id: Id945750af09913caafcfc198fcbc6e9cb32bbd9a