Support passing a range to the count: option
Via the case equality operator.

Because sometimes test data is non-deterministic, even when we don't
want it to be.

This also makes the count: argument a possible extension point,
because while previously the comparison was controlled by the query
count, which was not under the caller's control, now the comparison
method can be implemented on the argument.