Support `on_counted_query` callback
Extensive use of DBQueryMatchers can result in test assertions that are
hard to reason about. For example, the Brigade test suite has multiple
assertions that over 20, 30, or 40 queries are made. That's a lot to
debug if you're making some kind of sweeping change!

In order to provide users with a better time debugging/profiling exactly
which queries are being logged, this adds a configuration option that
accepts a proc to be called whenever a query is logged.

This is the simplest possible implementation. If people find this
useful, we can later support multiple callbacks, or some kind of better
API.