Check queries against primary or replica by role in Rails 6 (#35)
This change updates the query matchers to allow verification that
queries ran against the primary or replica database. This can be useful
in Rails 6 to verify the database role that was used, as test suites in
Rails by default group all database use under a single connection. The
role is still correctly set in test, allowing us to check that we made
queries against the correct role.