Add `database_role` option to scope checks for replicas
It can be useful to confirm that certain database queries are being made
against a primary database or a replica database. This change adds a new
option `database_role`, which makes the query counter only count queries
that were made against a database with the specified role.

We've used this logic at CommonLit for several years to ensure we're
querying the right database. This works in test environments even if
your primary and replica database configurations point at the same
database.